% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ATAC_normCount}
\alias{ATAC_normCount}
\title{ATAC normCount of E50h-72h in Chr5}
\format{
A matrix
}
\source{
\url{https://doi.org/10.1016/j.devcel.2020.07.003}
}
\usage{
data(ATAC_normCount)
}
\description{
ATAC normCount of E50h-72h in Chr5
}
\keyword{datasets}
