% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRP.R
\name{calcRP_region}
\alias{calcRP_region}
\title{calcRP_region}
\usage{
calcRP_region(
  mmAnno,
  peakScoreMt,
  Txdb,
  Chrs_included,
  decay_dist = 1000,
  log_transform = FALSE
)
}
\arguments{
\item{mmAnno}{the annotated GRange object from mm_geneScan}

\item{peakScoreMt}{peak count matrix. The rownames are feature_id in mmAnno,
while the colnames are sample names}

\item{Txdb}{Txdb}

\item{Chrs_included}{chromosome where you want to calcluate gene RP in}

\item{decay_dist}{decay distance}

\item{log_transform}{whether you want to log and norm your RP}
}
\value{
a MultiAssayExperiment object containg detailed peak-RP-gene relationship
and sumRP info
}
\description{
calculate regulatory potential based on mm_geneScan result and peakCount matrix,
which is useful for ATAC or H3K27ac histone modification data.
}
\examples{

data("ATAC_normCount")

library(TxDb.Athaliana.BioMart.plantsmart28)
library(SummarizedExperiment)
Txdb <- TxDb.Athaliana.BioMart.plantsmart28
seqlevels(Txdb) <- c(paste0("Chr", 1:5), "M", "C")
peak_path <- system.file("extdata", "ATAC.bed.gz", package = "FindIT2")
peak_GR <- loadPeakFile(peak_path)
mmAnno <- mm_geneScan(peak_GR,Txdb)

calcRP_region(mmAnno = mmAnno,
              peakScoreMt = ATAC_normCount,
              Txdb = Txdb,
              Chrs_included = "Chr5") -> regionRP

sumRP <- assays(regionRP)$sumRP
fullRP <- assays(regionRP)$fullRP


}
