% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRP.R
\name{calcRP_TFHit}
\alias{calcRP_TFHit}
\title{calcRP_TFHit}
\usage{
calcRP_TFHit(
  mmAnno,
  Txdb,
  decay_dist = 1000,
  consider_score = FALSE,
  report_fullInfo = FALSE
)
}
\arguments{
\item{mmAnno}{the annotated GRange object from mm_geneScan}

\item{Txdb}{Txdb}

\item{decay_dist}{decay distance}

\item{consider_score}{whether you want to consider score in your feature_score column}

\item{report_fullInfo}{whether you want to report full peak-RP-gene info}
}
\value{
if report_fullInfo is TRUE, it will output GRanges with detailed info.
While FALSE, it will output data frame
}
\description{
calculate regulatory potential based on ChIP-Seq peak data, which is useful
for TF ChIP-seq data.
}
\examples{
library(TxDb.Athaliana.BioMart.plantsmart28)
Txdb <- TxDb.Athaliana.BioMart.plantsmart28
seqlevels(Txdb) <- c(paste0("Chr", 1:5), "M", "C")
peak_path <- system.file("extdata", "ChIP.bed.gz", package = "FindIT2")
peak_GR <- loadPeakFile(peak_path)
mmAnno <- mm_geneScan(peak_GR,Txdb)

# if you just want to get RP_df, you can set report_fullInfo FALSE
fullRP_hit <- calcRP_TFHit(mmAnno = mmAnno,
                           Txdb = Txdb,
                           report_fullInfo = TRUE)

RP_df <- metadata(fullRP_hit)$peakRP_gene

}
