% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_influential_TF.R
\name{findIT_regionRP}
\alias{findIT_regionRP}
\title{findI(nfluential)T(F)_regionRP}
\usage{
findIT_regionRP(
  regionRP,
  Txdb,
  TF_GR_database,
  input_genes,
  background_genes = NULL,
  background_number = 3000
)
}
\arguments{
\item{regionRP}{the MultiAssayExperiment object from calcRP_region}

\item{Txdb}{Txdb}

\item{TF_GR_database}{TF peak GRange with a column named TF_id representing you TF name}

\item{input_genes}{genes which you want to find influentail TF for}

\item{background_genes}{If you do not assign background gene , program will sample
background_number genes as background genes from all gene sets.}

\item{background_number}{background genes number}
}
\value{
a MultiAssayExperiment object containg detailed TF-percent and TF-pvalue
}
\description{
find Influential TF of your input gene set based on regulatory potential data and TF ChIP-Seq or
motif data
}
\examples{
#see in vignettes

}
