% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{input_feature_id}
\alias{input_feature_id}
\title{input_feature_id}
\format{
feature_id vector
}
\usage{
data(input_feature_id)
}
\description{
input_feature_id
}
\details{
for the detailed progress producing input_feature_id, you can see data(input_genes)
}
\keyword{datasets}
