% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findIT_TTpair_enichAll_parse.R
\name{jaccard_findIT_TTpair}
\alias{jaccard_findIT_TTpair}
\title{jaccard_findIT_TTpair}
\usage{
jaccard_findIT_TTpair(input_genes, TF_target_database, input_TF_id)
}
\arguments{
\item{input_genes}{genes which you want to find influentail TF for
(same as your find_IT_TTpair parameter)}

\item{TF_target_database}{TF_target pair data}

\item{input_TF_id}{TF_id which you want to calculate jaccard index for}
}
\value{
jaccard similarity matrix
}
\description{
jaccard_findIT_TTpair
}
