% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RNA_normCount}
\alias{RNA_normCount}
\title{RNA normCount of E50h-72h in Chr5}
\format{
A matrix
}
\source{
\url{https://doi.org/10.1016/j.devcel.2020.07.003}
}
\usage{
data(RNA_normCount)
}
\description{
RNA normCount of E50h-72h in Chr5
}
\keyword{datasets}
