% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RNADiff_LEC2_GR}
\alias{RNADiff_LEC2_GR}
\title{RNA diff result from LEC2_GR VS LEC2_DMSO}
\format{
a data frame
}
\source{
\url{https://doi.org/10.1016/j.devcel.2020.07.003}
}
\usage{
data(RNADiff_LEC2_GR)
}
\description{
RNA diff result from LEC2_GR VS LEC2_DMSO
}
\keyword{datasets}
