% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findIT_regionRP_parse.R
\name{shinyParse_findIT_regionRP}
\alias{shinyParse_findIT_regionRP}
\title{shinyParse_findIT_regionRP}
\usage{
shinyParse_findIT_regionRP(merge_result, mode = c("gene", "TF"))
}
\arguments{
\item{merge_result}{the merge result from c(regionRP, TF_result)}

\item{mode}{one of 'gene' or 'TF'}
}
\value{
shiny website
}
\description{
if you have regionRP result from calcRP_region, and TF_result from
findIT_regionRP. you can just c(regionRP, TF_result) to merge two result. Then
use shinyParse_findIT_regionRP to parse the result
}
