% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakGeneCor_parse.R
\name{shinyParse_peakGeneCor}
\alias{shinyParse_peakGeneCor}
\title{shinyParse_peakGeneCor}
\usage{
shinyParse_peakGeneCor(mmAnnoCor)
}
\arguments{
\item{mmAnnoCor}{the annotated GRange object from peakGeneCor}
}
\value{
shiny mode
}
\description{
shinyParse_peakGeneCor
}
\examples{
if (require(TxDb.Athaliana.BioMart.plantsmart28)) {
    data("RNA_normCount")
    data("ATAC_normCount")
    Txdb <- TxDb.Athaliana.BioMart.plantsmart28
    seqlevels(Txdb) <- paste0("Chr", c(1:5, "M", "C"))
    peak_path <- system.file("extdata", "ATAC.bed.gz", package = "FindIT2")
    peak_GR <- loadPeakFile(peak_path)[1:100]
    mmAnno <- mm_geneScan(peak_GR, Txdb)

    ATAC_colData <- data.frame(
        row.names = colnames(ATAC_normCount),
        type = gsub("_R[0-9]", "", colnames(ATAC_normCount))
    )

    ATAC_normCount_merge <- integrate_replicates(ATAC_normCount, ATAC_colData)
    RNA_colData <- data.frame(
        row.names = colnames(RNA_normCount),
        type = gsub("_R[0-9]", "", colnames(RNA_normCount))
    )
    RNA_normCount_merge <- integrate_replicates(RNA_normCount, RNA_colData)
    mmAnnoCor <- peakGeneCor(
        mmAnno = mmAnno,
        peakScoreMt = ATAC_normCount_merge,
        geneScoreMt = RNA_normCount_merge,
        parallel = FALSE
    )

    # shinyParse_peakGeneCor(mmAnnoCor)

}
}
