% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_featureSet}
\alias{test_featureSet}
\title{test_featureSet}
\format{
character vector represent your interesting feature_id set
}
\usage{
data(test_featureSet)
}
\description{
test_featureSet
}
\details{
For the detailed progress producing input_feature_id, you can see ?test_geneSet
}
\keyword{datasets}
