% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TF_target_database}
\alias{TF_target_database}
\title{TF-target database}
\format{
a data frame
}
\source{
\url{http://bioinformatics.psb.ugent.be/webtools/iGRN/pages/download}
}
\usage{
data(TF_target_database)
}
\description{
TF-target database
}
\keyword{datasets}
