% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_influential_TF.R
\name{findIT_enrichInAll}
\alias{findIT_enrichInAll}
\title{findI(nfluential)T(F)_enrichInAll}
\usage{
findIT_enrichInAll(input_feature_id, peak_GR, TF_GR_database)
}
\arguments{
\item{input_feature_id}{peaks which you want to find influentail TF for}

\item{peak_GR}{all peak sets.Your input_feature_id is a part of it.}

\item{TF_GR_database}{TF peak GRange with a column named TF_id representing you TF name}
}
\value{
data.frame
}
\description{
find influential TF of your input peak set compared with your whole peak sets
based on TF ChIP-Seq or motif data.
}
\examples{
data("input_feature_id")
peak_path <- system.file("extdata", "ATAC.bed.gz", package = "FindIT2")
peak_GR <- loadPeakFile(peak_path)
ChIP_peak_path <- system.file("extdata", "ChIP.bed.gz", package = "FindIT2")
ChIP_peak_GR <- loadPeakFile(ChIP_peak_path)
ChIP_peak_GR$TF_id <- "AT1G28300"

findIT_enrichInAll(input_feature_id = input_feature_id,
peak_GR = peak_GR,
TF_GR_database = ChIP_peak_GR) -> result_findIT_enrichInAll


}
