% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_influential_TF.R
\name{findIT_MARA}
\alias{findIT_MARA}
\title{findIT_MARA}
\usage{
findIT_MARA(
  input_feature_id,
  peak_GR,
  peakScoreMt,
  TF_GR_database,
  log = TRUE,
  meanScale = TRUE,
  output = "coef"
)
}
\arguments{
\item{input_feature_id}{peaks which you want to find influentail TF for}

\item{peak_GR}{all peak sets.Your input_feature_id is a part of it.}

\item{peakScoreMt}{peak count matrix.}

\item{TF_GR_database}{TF peak GRange with a column named TF_id representing you TF name}

\item{log}{whether you want to log your peakScoreMt}

\item{meanScale}{whether you want to mean-centered per row}

\item{output}{one of 'coef' and 'cor'. Default is coef}
}
\value{
a data.frame
}
\description{
findIT_MARA
}
\examples{

data("ATAC_normCount")
data("input_feature_id")

peak_path <- system.file("extdata", "ATAC.bed.gz", package = "FindIT2")
peak_GR <- loadPeakFile(peak_path)

ChIP_peak_path <- system.file("extdata", "ChIP.bed.gz", package = "FindIT2")
ChIP_peak_GR <- loadPeakFile(ChIP_peak_path)
ChIP_peak_GR$TF_id <- "AT1G28300"

set.seed(20160806)

findIT_MARA(
    input_feature_id = input_feature_id,
    peak_GR = peak_GR,
    peakScoreMt = ATAC_normCount,
    TF_GR_database = ChIP_peak_GR
) -> result_findIT_MARA

}
