% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_influential_TF.R
\name{findIT_TFHit}
\alias{findIT_TFHit}
\title{findI(nfluential)T(F)_TFHit}
\usage{
findIT_TFHit(
  input_genes,
  Txdb,
  TF_GR_database,
  scan_dist = 1000,
  decay_dist = 20000,
  Chrs_included,
  background_genes = NULL,
  background_number = 3000
)
}
\arguments{
\item{input_genes}{genes which you want to find influentail TF for}

\item{Txdb}{Txdb}

\item{TF_GR_database}{TF peak GRange with a column named TF_id representing you TF name}

\item{scan_dist}{scan distance}

\item{decay_dist}{decay distance}

\item{Chrs_included}{chromsomes which you want to sample background genes from}

\item{background_genes}{If you do not assign background gene , program will sample
background_number genes as background genes from all gene sets.}

\item{background_number}{background genes number}
}
\value{
data.frame
}
\description{
find influential TF of your input gene set based on TF ChIP-Seq or motif data
}
\examples{
#see in vignettes

}
