% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findIT_TTpair_enichAll_parse.R
\name{jaccard_findIT_enrichInAll}
\alias{jaccard_findIT_enrichInAll}
\title{jaccard_findIT_enrichInAll}
\usage{
jaccard_findIT_enrichInAll(
  input_feature_id,
  peak_GR,
  TF_GR_database,
  input_TF_id
)
}
\arguments{
\item{input_feature_id}{peaks which you want to find influentail TF for
(same as your find_IT_enrichInAll parameter)}

\item{peak_GR}{all peak sets.Your input_feature_id is a part of it.}

\item{TF_GR_database}{TF peak GRange with a column named TF_id representing you TF name}

\item{input_TF_id}{TF_id which you want to calculate jaccard index for}
}
\value{

}
\description{
jaccard_findIT_enrichInAll
}
