% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{loadPeakFile}
\alias{loadPeakFile}
\title{loadPeakFile}
\usage{
loadPeakFile(filePath, fromMACS2 = FALSE, narrowPeak = TRUE)
}
\arguments{
\item{filePath}{peak Path}

\item{fromMACS2}{whether peak from MACS2.If it set TRUE, extra column in bed
file will be named according to the content in MACS2 manual}

\item{narrowPeak}{whether peak is narrowPeak or broadPeak.}
}
\value{
GRanges object with a column named feature_id, which is important
for later analysis
}
\description{
read peak file and transform it into GRanges object
}
\examples{
peakfile <- system.file("extdata", "ChIP.bed.gz", package = "MPMG")
loadPeakFile(peakfile)
}
