% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakGeneCor.R
\name{peakGeneCor}
\alias{peakGeneCor}
\title{peakGeneCor}
\usage{
peakGeneCor(mmAnno, peakScoreMt, geneScoreMt, parallel = FALSE)
}
\arguments{
\item{mmAnno}{the annotated GRange object from mm_geneScan or mm_nearestGene}

\item{peakScoreMt}{peak count matrix. The rownames are feature_id in mmAnno,
while the colnames are sample names.}

\item{geneScoreMt}{gene count matirx. The rownames are gene_id in mmAnno,
while the colnames are sample names.}

\item{parallel}{whehter you want to using bplapply to speed up calculation}
}
\value{
mmAnno with Cor, pvalue,padj,qvalue column
}
\description{
peakGeneCor
}
\examples{
data("RNA_normCount")
data("ATAC_normCount")

library(TxDb.Athaliana.BioMart.plantsmart28)
Txdb <- TxDb.Athaliana.BioMart.plantsmart28
seqlevels(Txdb) <- c(paste0("Chr", 1:5), "M", "C")
peak_path <- system.file("extdata", "ATAC.bed.gz", package = "FindIT2")
peak_GR <- loadPeakFile(peak_path)[1:100]
mmAnno <- mm_geneScan(peak_GR,Txdb)

ATAC_colData <- data.frame(row.names = colnames(ATAC_normCount),
                           type = gsub("_R[0-9]", "", colnames(ATAC_normCount))
                           )

integrate_replicates(ATAC_normCount, ATAC_colData) -> ATAC_normCount_merge
RNA_colData <- data.frame(row.names = colnames(RNA_normCount),
                          type = gsub("_R[0-9]", "", colnames(RNA_normCount))
                          )
integrate_replicates(RNA_normCount, RNA_colData) -> RNA_normCount_merge
peakGeneCor(mmAnno = mmAnno,
            peakScoreMt = ATAC_normCount_merge,
            geneScoreMt = RNA_normCount_merge,
            parallel = FALSE) -> mmAnnoCor

mmAnnoCor

}
