% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TF_target_database}
\alias{TF_target_database}
\title{TF-target database}
\format{
a data frame
}
\source{
\url{http://bioinformatics.psb.ugent.be/webtools/iGRN/pages/download}
}
\usage{
data(TF_target_database)
}
\description{
TF-target database
}
\examples{
\dontrun{
# source
library(dplyr)
data <- read.table("~/reference/annoation/Athaliana/TF_target/iGRN_network_full.txt",
                  sep = "\t",
                  stringsAsFactors = FALSE)

data \%>\%
rename(TF_id = V1, target_gene = V2) \%>\%
select(TF_id, target_gene) \%>\%
filter(TF_id \%in\% c("AT1G28300", "AT5G63790", "AT5G24110", "AT3G23250")) \%>\%
as.data.frame() -> TF_target_database

save(TF_target_database, file = "inst/extdata/TF_target_database.rda", version = 2,
     compress = "bzip2")

}

}
\keyword{datasets}
